WIP: Attempt to install Rack::MiniProfiler
## Problem

Our application is slow,
and we have little visibility into what is taking the most time
in each request.

## Solution

We attempted to profile our application using the process described in:
https://www.nateberkopec.com/2015/08/05/rack-mini-profiler-the-secret-weapon.html

but we could not get `rack-mini-profiler` set up
in our production environment.

The gem is not as useful in development,
because the page load times and SQL queries
are vastly different from production.
(See http://stackoverflow.com/questions/14694392#14694554)

Until we can get the gem working in our production environment,
we will keep it in a separate branch.